+++ src/graphql/outlier.rs
@@ -1,4 +1,4 @@
-use super::{always_true, model::Model, Role, RoleGuard, DEFAULT_CONNECTION_SIZE};
+use super::{always_true, model::ModelDigest, Role, RoleGuard, DEFAULT_CONNECTION_SIZE};
 use anyhow::anyhow;
 use async_graphql::{
     connection::{query, Connection, Edge, EmptyFields},
@@ -138,7 +138,7 @@ async fn load_outliers(
     filter: fn(RankedOutlier) -> Option<RankedOutlier>,
 ) -> Result<Connection<String, RankedOutlier, RankedOutlierTotalCount, EmptyFields>> {
     let model_id: i32 = model_id.as_str().parse()?;
-    let timestamp = time.map(|t| t.timestamp_nanos());
+    let timestamp = time.map(|t| t.timestamp_nanos_opt().unwrap_or_default());
 
     let prefix = if let Some(timestamp) = timestamp {
         bincode::DefaultOptions::new().serialize(&(model_id, timestamp))?
@@ -307,7 +307,7 @@ impl Outlier {
             .collect::<Vec<_>>())
     }
 
-    async fn model(&self, ctx: &Context<'_>) -> Result<Model> {
+    async fn model(&self, ctx: &Context<'_>) -> Result<ModelDigest> {
         let db = ctx.data::<Database>()?;
         Ok(db.load_model(self.model_id).await?.into())
     }